+++ src/class/view-item.ts
@@ -5,6 +5,8 @@ import * as vscode from "vscode";
 import { ResourceType } from "../types/index";
 import { Favorites } from "./favorites";
 
+export type ViewItemContextValue = "FAVORITE_GROUP" | "FAVORITE_DIRECTORY" | "FAVORITE_FILE" | "FS_DIRECTORY" | "FS_FILE";
+
 export class ViewItem extends vscode.TreeItem {
     public static favorites: Favorites;
     public static context: vscode.ExtensionContext;
@@ -17,7 +19,7 @@ export class ViewItem extends vscode.TreeItem {
         public label: string,
         public collapsibleState: vscode.TreeItemCollapsibleState,
         public value: string,
-        public contextValue: string,
+        public contextValue: ViewItemContextValue,
         public resourceName: string,
         public resourceType: ResourceType,
         public icon?: string | vscode.Uri | { light: string | vscode.Uri; dark: string | vscode.Uri },
@@ -37,7 +39,6 @@ export class ViewItem extends vscode.TreeItem {
         return this.contextValue === "FAVORITE_DIRECTORY" || this.contextValue === "FAVORITE_FILE";
     }
     public getParent(): Promise<ViewItem> {
-        // return new Promise((resolve, reject) => {
 
         if (
             this.contextValue === "FAVORITE_DIRECTORY" ||
@@ -52,11 +53,40 @@ export class ViewItem extends vscode.TreeItem {
             return this.getParentForFavorite(this.parentId);
 
         }
-        if (this.contextValue === "FS_FILE" || this.contextValue === "FS_DIRECTORY") {
 
-            return this.getParentForFs(this.resourceName);
-        }
+        return new Promise((resolve, reject) => {
 
+            ViewItem.favorites.get()
+                .then((stored) => {
+
+                    this.getParentForFs(this.resourceUri.fsPath)
+                        .then((result) => {
+
+                            return Promise.all([
+                                ViewItem.favorites.isPartOfFavorites(result.resourceUri.fsPath),
+                                result,
+                            ]);
+
+                        }).then((res) => {
+
+                            const check = res[0];
+                            const item = res[1];
+
+                            if (check) {
+                                resolve(item);
+                            } else {
+                                resolve(null);
+                            }
+
+                        }).catch((e) => {
+                            reject(e);
+                        });
+
+                }).catch((e) => {
+                    reject(e);
+                });
+
+        });
     }
     private getParentForFs(fsPath: string): Promise<ViewItem> {
         return new Promise((resolve, reject) => {